requirements.txt
mysql-connector-python>=8.0
pandas>=1.3
requests>=2.25
python-dotenv>=1.0
gspread>=5.0.0
oauth2client>=4.1.3
